Added wrappers for Volume, VolumetricModel and TransferFunction
Added conversion of 3-dimensional NumPy arrays to Data objects.

Added conversion of NumPy arrays of shape (N,1), for example to
pass TF opacties (which need to be a float[] value).

Added some wrapping for ospcommon::math::affine3f

OSPDataType enums where not exported in the top namespace

Allow string parameters to be set